rename: lww 파일 이름에서 빼기
